Change delim between hdr length and hdr to \r\n
There are builtin facilities in most languages for splitting on lines in a buffered file or network IO situation. These basically all support `\r\n` or `\n` as the newline ending, but many don't support just `\r`, because a common newline regex is `\r?\n`, which will match an optional `\r` followed by a mandatory `\n`. `\r\n` is the HTTP standard newline and although the streaming code in this file does new use HTTP, other sensors that POST over HTTP will have to use the `\r\n` newline, so let's standardize on it for cleartext data over the network.

This change is compatible with the current streaming server implementation and the exact format for streaming data is only "documented" in this reference implementation AFAIK. Please point me to any documentation that needs to be updated.

https://github.com/usnistgov/SpectrumBrowser/blob/master/services/streaming/StreamingServer.py#L314